RWA-3584 changes
