--- src/integrationTest/resources/scripts/wa/delete_task_data.sql
@@ -1,2 +1,0 @@
-DELETE FROM cft_task_db.tasks where task_id ='9a6cc5cf-c973-11eb-bdba-0242ac111001';
-DELETE FROM cft_task_db.tasks where task_id ='9a6cc5cf-c973-11eb-bdba-0242ac111011';
