--- src/main/resources/db/migration/V1.0.33__termination_process_enum_plsql.sql
@@ -1,5 +1,0 @@
-DROP TYPE IF EXISTS termination_process_enum;
-CREATE TYPE termination_process_enum as ENUM (
-    'EXUI_USER_COMPLETION',
-    'EXUI_CASE-EVENT_COMPLETION');
-COMMIT;
