--- src/main/resources/db/migration/V1.0.34__add_column_termination_process_plsql.sql
@@ -1,1 +1,0 @@
-ALTER TABLE cft_task_db.tasks ADD COLUMN termination_process termination_process_enum;
